looking for responce in postman dated 1/18 5:00

+++ springbootRestfulWebService/src/main/java/com/example/entity/AddTasks.java
@@ -16,7 +16,7 @@ public class AddTasks {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long TaskId;
+    private Integer TaskId;
     private String TaskName;
     private String TaskDescription;
     private String TaskDate;
